randevents: do not cast timestamp in integer
Otherwise, we lose precision and this causes trouble with insertion in
InfluxDB because two points of the same series cannot have the same
timestamp.

+++ loudml/loudml/randevents.py
@@ -69,7 +69,7 @@ class EventGenerator(metaclass=ABCMeta):
             nb_events = int(nb_events) + extra
 
             for i in range(nb_events):
-                yield int(ts + i * step / nb_events)
+                yield ts + i * step / nb_events
 
 
 class FlatEventGenerator(EventGenerator):
